+++ src/controller/publications/PublicationController.ts
@@ -21,11 +21,11 @@ export class PublicationController {
 
   async all(request: Request, response: Response, next: NextFunction) {
     const publications = await this.publicationRepository.find();
-    return publications
+    return publications;
   }
 
   async remove(request: Request, response: Response, next: NextFunction) {
-    const slug = request.params.slug
+    const slug = request.params.slug;
 
     const publicationToRemove = await this.publicationRepository.findOne({ where: { slug } });
 
@@ -36,5 +36,17 @@ export class PublicationController {
     await this.publicationRepository.remove(publicationToRemove);
 
     return "La Publicación se ha borrado correctamente";
+  }
+
+  async update(request: Request, response: Response, next: NextFunction) {
+    const slug = request.params.slug;
+    const publication = await this.publicationRepository.findOne({ where: { slug } });
+    if (!publication) { return `No se encontró publicación`; }
+    const { name, initialContent, finalContent } = request.body;
+    publication.name = name;
+    publication.initialContent = initialContent;
+    publication.finalContent = finalContent;
+    await this.publicationRepository.save(publication);
+    return publication;
   }
 }
